LibCore: Added basic IMAPClient class that can connect to an IMAP server
Right now, the IMAPClient class is in LibCore, with the ability to
connect to an IMAP server using a given IP address and port. While
it is able to make the initial connection to a server, there seems
to be some trouble in the TCP library that is causing a seg fault.